Marker caused null pointers - fixed. Sending only acceptable files to compilation. Additionally, bringing problem marker back to buildpath.txt. Unfortunately, eclipse caches the image decoration and we can see BP file error only in status bar. That is,
class eclipse.ui.internal.decorators.DecorationScheduler

	private DecorationResult internalGetResult(Object element,
			IDecorationContext context) {
		Map results = (Map) resultCache.get(context);

returns null for all normally Java Project Explorer deocrated file images. But, some weired result is cached after status bar icon is derived for buildpath. So that you see the problem in status bar but not in Navigator.

This is because internalGetResult return null in

	private DecorationResult getResult(Object element, Object adaptedElement,
			IDecorationContext context) {

		// We do not support decoration of null
		if (element == null) {
			return null;
		}

		DecorationResult decoration = internalGetResult(element, context);

		if (decoration == null) {
			queueForDecoration(element, adaptedElement, false, null, context);
			return null;
		}

and queueForDecoration creates the decoration cache job in case of buildpath. Curiously, cache is zero when I select the BuildPath icon with mouse but it reappears again after status bar.
In case of vhdl, a job is started to add the chacked decoration, but it does not reach the DecorationScheduler.internalPutResult

On once case, cache succeeds because

		contentDescription = file.getContentDescription();
		if (contentDescription != null) {
			IContentType contentType = contentDescription.getContentType();
			if (contentType != null) {
				image = workbench.getEditorRegistry().getImageDescriptor(file.getName(), contentType);
			}

returns the BuildPath image in ContentTypeDecorator.decorate()
because contentDescription = default org.eclipse.core.runtime.text. It is null in other cases because vhdl files have M_NO_CONTENT_DESCRIPTION attribute.
Clearly, java label provider calls something like image = workbench.getEditorRegistry().getImageDescriptor(file.getName(), contentType);
But, it also adds decoration, which is missing here.

Described the same in http://www.eclipse.org/forums/index.php/m/779463/#msg_779463

+++ src/org/zamia/plugin/build/ZamiaErrorObserver.java
@@ -110,7 +110,7 @@ public class ZamiaErrorObserver implements ErrorObserver {
 
 		IFile file = ZamiaPlugin.getIFile(aSF, fProject);
 
-		if (file != null) {
+		if (file != null && file.exists()) { // file will not exist if it was renamed. By the time this called, errors are removed from ZDB. But, I do not know about eclipse markers, linked to the old file name. How to remove them?
 
 			try {
 				file.deleteMarkers(null, true, IResource.DEPTH_ZERO);
